+++ src/jsdoc-globals.js
@@ -180,6 +180,7 @@ var FancytreeOptions = {};
  * @property {function} expand `data.node` was expanded
  * @property {function} focus `data.node` received keyboard focus
  * @property {function} focusTree `data.tree` received keyboard focus
+ * @property {function} iconClass Allows to override the class names added to the `data.node icon markup to define custom icons or glyphs (separate with space, e.g. `ui-icon ui-icon-heart`).
  * @property {function} init Widget was (re-)initialized.
  * @property {function} keydown `data.node` received key. `event.which` contains the key. Return `false` to prevent default processing, i.e. navigation. Call `data.result = "preventNav";` to prevent navigation but still allow default handling inside embedded input controls.
  * @property {function} keypress (currently unused)